Filter by flow if navigating from flow
